Refactor UI: move all render functions to src/ui module
- Moved draw_timer, draw_task_list, draw_statistics, draw_task_details,
  and
  draw_settings into src/ui/
- Split settings and theme logic into src/settings/
- Removed old theme.rs and settings.rs, replaced with modularized
  versions
- Updated tasks.md to mark UI split as complete
