+++ src/app/mod.rs
@@ -1,4 +1,4 @@
-use crate::settings::{ColorTheme, Settings};
+use crate::settings::Settings;
 use chrono::{DateTime, Utc};
 use directories::ProjectDirs;
 use serde::{Deserialize, Serialize};
@@ -6,7 +6,8 @@ use std::fs;
 use std::path::PathBuf;
 use std::time::Duration;
 
-const SETTINGS_ROW_COUNT: usize = 6;
+pub mod ui_state;
+pub use ui_state::UiState;
 
 fn project_dirs() -> Option<ProjectDirs> {
     ProjectDirs::from("", "", "pomodorust")
@@ -20,7 +21,6 @@ pub fn get_config_path() -> Option<PathBuf> {
     project_dirs().map(|d| d.config_dir().join("config.toml"))
 }
 
-/// Represents a single task for the Pomodoro timer.
 #[derive(Clone, Serialize, Deserialize, Debug)]
 pub struct Task {
     pub name: String,
@@ -32,7 +32,6 @@ pub struct Task {
 }
 
 impl Task {
-    /// Creates a new task with a given name.
     pub fn new(name: String) -> Self {
         Self {
             name,
@@ -45,7 +44,6 @@ impl Task {
     }
 }
 
-/// Represents the different timer modes in the Pomodoro technique.
 #[derive(Clone, Copy, PartialEq, Serialize, Deserialize, Default)]
 pub enum Mode {
     #[default]
@@ -55,7 +53,6 @@ pub enum Mode {
 }
 
 impl Mode {
-    /// Returns the duration of the timer mode based on settings.
     pub fn duration(&self, settings: &Settings) -> Duration {
         match self {
             Mode::Pomodoro => settings.pomodoro_duration,
@@ -64,7 +61,6 @@ impl Mode {
         }
     }
 
-    /// Returns the title of the timer mode.
     pub fn title(&self) -> &'static str {
         match self {
             Mode::Pomodoro => "Pomodoro",
@@ -74,7 +70,6 @@ impl Mode {
     }
 }
 
-/// Represents the current state of the timer.
 #[derive(Clone, Copy, PartialEq, Serialize, Deserialize, Default)]
 pub enum TimerState {
     #[default]
@@ -82,7 +77,6 @@ pub enum TimerState {
     Running,
 }
 
-/// Represents the different views of the application.
 #[derive(Serialize, Deserialize, Default, PartialEq, Eq, Clone, Copy, Debug)]
 pub enum View {
     Timer,
@@ -93,7 +87,6 @@ pub enum View {
     TaskDetails,
 }
 
-/// Represents the different input modes.
 #[derive(Default)]
 pub enum InputMode {
     #[default]
@@ -101,7 +94,6 @@ pub enum InputMode {
     Editing,
 }
 
-/// The main application state (business logic + persistence).
 #[derive(Serialize, Deserialize)]
 #[serde(default)]
 pub struct App {
@@ -118,7 +110,7 @@ pub struct App {
     pub settings: Settings,
 }
 
-fn bump_duration_mins(d: Duration, delta: i64) -> Duration {
+pub(super) fn bump_duration_mins(d: Duration, delta: i64) -> Duration {
     let mins = (d.as_secs() / 60) as i64;
     Duration::from_secs((mins + delta).max(1) as u64 * 60)
 }
@@ -141,7 +133,6 @@ impl Default for App {
 }
 
 impl App {
-    /// Loads an App instance from a file, or creates a new one.
     pub fn load_with_settings(settings: Settings) -> Self {
         let mut app: App = if let Some(path) = get_data_path() {
             fs::read_to_string(path)
@@ -151,15 +142,12 @@ impl App {
         } else {
             App::default()
         };
-
         app.settings = settings;
         app.time_remaining = app.mode.duration(&app.settings);
         app
     }
 
-    /// Saves the current state of the app to a file.
     pub fn save(&self) {
-        // Save the main app state (tasks, etc.)
         if let Some(path) = get_data_path() {
             if let Some(parent) = path.parent() {
                 if fs::create_dir_all(parent).is_ok() {
@@ -169,11 +157,9 @@ impl App {
                 }
             }
         }
-        // Save the settings
         self.settings.save();
     }
 
-    /// Toggles the timer between running and paused states.
     pub fn toggle_timer(&mut self) {
         if let Some(index) = self.active_task_index {
             if !self.tasks[index].completed {
@@ -185,13 +171,11 @@ impl App {
         }
     }
 
-    /// Resets the timer to the current mode's full duration.
     pub fn reset_timer(&mut self) {
         self.state = TimerState::Paused;
         self.time_remaining = self.mode.duration(&self.settings);
     }
 
-    /// Switches to the next appropriate timer mode.
     pub fn next_mode(&mut self) -> Mode {
         let previous_mode = self.mode;
         if self.mode == Mode::Pomodoro {
@@ -201,7 +185,6 @@ impl App {
                     task.pomodoros += 1;
                 }
             }
-
             let interval = self.settings.long_break_interval.max(1) as u32;
             if self.pomodoros_completed_total % interval == 0 {
                 self.mode = Mode::LongBreak;
@@ -220,14 +203,11 @@ impl App {
         previous_mode
     }
 
-    /// Sets the current mode explicitly.
     pub fn set_mode(&mut self, mode: Mode) {
         self.mode = mode;
         self.reset_timer();
     }
 
-
-    /// Toggles the completion status of the active task.
     pub fn complete_active_task(&mut self) {
         if let Some(index) = self.active_task_index {
             if let Some(task) = self.tasks.get_mut(index) {
@@ -246,7 +226,6 @@ impl App {
         }
     }
 
-    /// Deletes the currently selected active (non-completed) task.
     pub fn delete_active_task(&mut self) {
         if let Some(index) = self.active_task_index {
             self.tasks.remove(index);
@@ -258,60 +237,30 @@ impl App {
         }
     }
 
-    /// Moves the selection to the next uncompleted task.
     pub fn next_task(&mut self) {
-        let uncompleted_tasks_indices: Vec<usize> = self
-            .tasks
-            .iter()
-            .enumerate()
+        let indices: Vec<usize> = self.tasks.iter().enumerate()
             .filter(|(_, t)| !t.completed)
             .map(|(i, _)| i)
             .collect();
-
-        if uncompleted_tasks_indices.is_empty() {
-            self.active_task_index = None;
-            return;
-        }
-
-        let current_selection = self.active_task_index.unwrap_or(0);
-        let next_index_in_uncompleted = uncompleted_tasks_indices
-            .iter()
-            .position(|&i| i == current_selection)
-            .map_or(0, |i| (i + 1) % uncompleted_tasks_indices.len());
-
-        self.active_task_index = Some(uncompleted_tasks_indices[next_index_in_uncompleted]);
+        if indices.is_empty() { self.active_task_index = None; return; }
+        let cur = self.active_task_index.unwrap_or(0);
+        let next = indices.iter().position(|&i| i == cur)
+            .map_or(0, |p| (p + 1) % indices.len());
+        self.active_task_index = Some(indices[next]);
     }
 
-    /// Moves the selection to the previous uncompleted task.
     pub fn previous_task(&mut self) {
-        let uncompleted_tasks_indices: Vec<usize> = self
-            .tasks
-            .iter()
-            .enumerate()
+        let indices: Vec<usize> = self.tasks.iter().enumerate()
             .filter(|(_, t)| !t.completed)
             .map(|(i, _)| i)
             .collect();
-
-        if uncompleted_tasks_indices.is_empty() {
-            self.active_task_index = None;
-            return;
-        }
-
-        let current_selection = self.active_task_index.unwrap_or(0);
-        let current_pos = uncompleted_tasks_indices
-            .iter()
-            .position(|&i| i == current_selection)
-            .unwrap_or(0);
-        let next_index_in_uncompleted = if current_pos == 0 {
-            uncompleted_tasks_indices.len() - 1
-        } else {
-            current_pos - 1
-        };
-
-        self.active_task_index = Some(uncompleted_tasks_indices[next_index_in_uncompleted]);
+        if indices.is_empty() { self.active_task_index = None; return; }
+        let cur = self.active_task_index.unwrap_or(0);
+        let pos = indices.iter().position(|&i| i == cur).unwrap_or(0);
+        let prev = if pos == 0 { indices.len() - 1 } else { pos - 1 };
+        self.active_task_index = Some(indices[prev]);
     }
 
-    /// Moves the currently active task up in the list.
     pub fn move_active_task_up(&mut self) {
         if let Some(index) = self.active_task_index {
             if index > 0 {
@@ -321,7 +270,6 @@ impl App {
         }
     }
 
-    /// Moves the currently active task down in the list.
     pub fn move_active_task_down(&mut self) {
         if let Some(index) = self.active_task_index {
             if index < self.tasks.len() - 1 {
@@ -329,111 +277,5 @@ impl App {
                 self.active_task_index = Some(index + 1);
             }
         }
-    }
-}
-
-/// Transient UI navigation state — not persisted.
-pub struct UiState {
-    pub settings_selection: usize,
-    pub completed_task_list_state: Option<usize>,
-    pub previous_view: View,
-    pub input_mode: InputMode,
-    pub current_input: String,
-}
-
-impl Default for UiState {
-    fn default() -> Self {
-        Self {
-            settings_selection: 0,
-            completed_task_list_state: None,
-            previous_view: View::TaskList,
-            input_mode: InputMode::Normal,
-            current_input: String::new(),
-        }
-    }
-}
-
-impl UiState {
-    pub fn next_setting(&mut self) {
-        self.settings_selection = (self.settings_selection + 1) % SETTINGS_ROW_COUNT;
-    }
-
-    pub fn previous_setting(&mut self) {
-        if self.settings_selection > 0 {
-            self.settings_selection -= 1;
-        } else {
-            self.settings_selection = SETTINGS_ROW_COUNT - 1;
-        }
-    }
-
-    pub fn modify_setting(&mut self, app: &mut App, increase: bool) {
-        let delta: i64 = if increase { 1 } else { -1 };
-        match self.settings_selection {
-            0 => app.settings.pomodoro_duration = bump_duration_mins(app.settings.pomodoro_duration, delta),
-            1 => app.settings.short_break_duration = bump_duration_mins(app.settings.short_break_duration, delta),
-            2 => app.settings.long_break_duration = bump_duration_mins(app.settings.long_break_duration, delta),
-            3 => {
-                app.settings.theme = match app.settings.theme {
-                    ColorTheme::Default => ColorTheme::Dracula,
-                    ColorTheme::Dracula => ColorTheme::Solarized,
-                    ColorTheme::Solarized => ColorTheme::Nord,
-                    ColorTheme::Nord => ColorTheme::Default,
-                };
-            }
-            4 => app.settings.desktop_notifications = !app.settings.desktop_notifications,
-            5 => {
-                let current = app.settings.long_break_interval as i64;
-                app.settings.long_break_interval = (current + delta).max(1) as u32;
-            }
-            _ => {}
-        }
-        if app.state == TimerState::Paused {
-            app.reset_timer();
-        }
-    }
-
-    pub fn next_completed_task(&mut self, app: &App) {
-        let count = app.tasks.iter().filter(|t| t.completed).count();
-        if count == 0 { return; }
-        let i = self.completed_task_list_state.map_or(0, |i| (i + 1) % count);
-        self.completed_task_list_state = Some(i);
-    }
-
-    pub fn previous_completed_task(&mut self, app: &App) {
-        let count = app.tasks.iter().filter(|t| t.completed).count();
-        if count == 0 { return; }
-        let i = self.completed_task_list_state.map_or(0, |i| {
-            if i == 0 { count - 1 } else { i - 1 }
-        });
-        self.completed_task_list_state = Some(i);
-    }
-
-    pub fn delete_selected_completed_task(&mut self, app: &mut App) {
-        if let Some(selected) = self.completed_task_list_state {
-            let completed_indices: Vec<usize> = app.tasks.iter().enumerate()
-                .filter(|(_, t)| t.completed)
-                .map(|(i, _)| i)
-                .collect();
-            if let Some(&idx) = completed_indices.get(selected) {
-                app.tasks.remove(idx);
-                if let Some(active) = app.active_task_index {
-                    if active > idx {
-                        app.active_task_index = Some(active - 1);
-                    }
-                }
-                self.completed_task_list_state = None;
-            }
-        }
-    }
-
-    pub fn submit_task(&mut self, app: &mut App) {
-        if !self.current_input.is_empty() {
-            app.tasks.push(Task::new(self.current_input.clone()));
-            self.current_input.clear();
-            if app.tasks.len() == 1 {
-                app.active_task_index = Some(0);
-            }
-        }
-        self.input_mode = InputMode::Normal;
     }
 }
