+++ src/ui/mod.rs
@@ -1,0 +1,11 @@
+pub mod details;
+pub mod settings;
+pub mod statistics;
+pub mod task_list;
+pub mod timer;
+
+pub use details::draw_task_details;
+pub use settings::draw_settings;
+pub use statistics::draw_statistics;
+pub use task_list::draw_task_list;
+pub use timer::draw_timer;
